requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
djangorestframework==3.13.1
Pillow==9.1.1
Pygments==2.12.0
python-decouple==3.6
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==37.0.2
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
jwt==1.3.1
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2

